requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-postgres>=0.0.12
langchain-community>=0.3.0
psycopg[binary]>=3.2.0
pypdf>=5.0.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
